fix(asr-worker): don't recreate symlink when it exists

+++ workers/asr-worker/asr_worker/activities.py
@@ -397,8 +397,9 @@ async def create_symlinks_for_embedded_audios(
             audio_ext = Path(d.resource_name).suffix
             symlink_path = d.path.relative_to(Path(d.index))
             symlink_path = symlinks_dir / f"{symlink_path}{audio_ext}"
-            symlink_path.parent.mkdir(parents=True, exist_ok=True)
-            os.symlink(artifact_path, symlink_path)
+            if not symlink_path.exists():
+                symlink_path.parent.mkdir(parents=True, exist_ok=True)
+                os.symlink(artifact_path, symlink_path)
             symlink = FilesystemDocument(
                 path=symlink_path.relative_to(workdir),
                 id=d.id,
